Acceptance tests - pull method up
* Moved the login method of MarketMemberTestBase to MarketTestBase
